ENH: outsourced SliceTrackerTrainingStep into separate python file
